app/requirements.txt
streamlit==1.28.1
psycopg2-binary==2.9.7
pandas==2.0.3
plotly==5.15.0